refactored setting proxy value
